Move mock data to central location
