+++ src/components/ShareExport/index.test.tsx
@@ -5,14 +5,11 @@ import { render, waitFor } from '@testing-library/react';
 import configureMockStore from 'redux-mock-store';
 import Sinon from 'sinon';
 
-import allMocks, { modelID, summaryMock } from 'data/mock/readonly';
-import GetOperationalNeeds from 'queries/ITSolutions/GetOperationalNeeds';
-import { GetOperationalNeeds as GetOperationalNeedsType } from 'queries/ITSolutions/types/GetOperationalNeeds';
-import {
-  OperationalNeedKey,
-  OperationalSolutionKey,
-  OpSolutionStatus
-} from 'types/graphql-global-types';
+import allMocks, {
+  modelID,
+  operationalNeedsMock,
+  summaryMock
+} from 'data/mock/readonly';
 import VerboseMockedProvider from 'utils/testing/MockedProvider';
 import setup from 'utils/testing/setup';
 
@@ -20,58 +17,6 @@ import ShareExportModal from './index';
 
 const mockStore = configureMockStore();
 const store = mockStore({ auth: { euaId: 'MINT' } });
-
-const opNeedsData: GetOperationalNeedsType = {
-  modelPlan: {
-    __typename: 'ModelPlan',
-    id: modelID,
-    isCollaborator: true,
-    modelName: 'My excellent plan that I just initiated',
-    operationalNeeds: [
-      {
-        __typename: 'OperationalNeed',
-        id: '123',
-        modelPlanID: modelID,
-        name: 'Recruit participants',
-        key: OperationalNeedKey.RECRUIT_PARTICIPANTS,
-        nameOther: null,
-        needed: true,
-        modifiedDts: '2022-05-12T15:01:39.190679Z',
-        solutions: [
-          {
-            __typename: 'OperationalSolution',
-            id: '123',
-            status: OpSolutionStatus.IN_PROGRESS,
-            name: 'Shared Systems',
-            key: OperationalSolutionKey.SHARED_SYSTEMS,
-            otherHeader: '',
-            mustStartDts: null,
-            mustFinishDts: null,
-            operationalSolutionSubtasks: [],
-            needed: true,
-            nameOther: null,
-            pocEmail: null,
-            pocName: null,
-            createdBy: '',
-            createdDts: ''
-          }
-        ]
-      }
-    ]
-  }
-};
-
-const operationalNeedMock = [
-  {
-    request: {
-      query: GetOperationalNeeds,
-      variables: { id: modelID }
-    },
-    result: {
-      data: opNeedsData
-    }
-  }
-];
 
 describe('ShareExportModal', () => {
   // Stubing Math.random that occurs in Truss Tooltip component for deterministic output
@@ -86,7 +31,7 @@ describe('ShareExportModal', () => {
           ]}
         >
           <VerboseMockedProvider
-            mocks={[...allMocks, ...summaryMock, ...operationalNeedMock]}
+            mocks={[...allMocks, ...summaryMock, ...operationalNeedsMock]}
             addTypename={false}
           >
             <Route path="/models/:modelID/read-only/model-basics">
